Add FFLD2 to menpodetect
This includes detection and fitting code. This is tightly coupled
with my custom changes to the ffld2 codebase, so this will
sit on a branch for now until I am sure I have managed to get
detection and training to work properly from cyffld2.

Also, some minor changes, in that the output detection keys
should now be prefixed with the detector type which helps with
overwriting the detections when you are testing the output
of multiple detectors.